Add emojis and Verified Dev
The emojis `_online_`, `_offline_`, `_away_`,  `_dnd_`, `_isonline_` are added.

On the isOnlineV2 profile, a dev member's comment is marked so there is no confusion that the person in question knows what they're talking about.

The style of this marker should be perhaps re-evaluated ;), but I think it's good enough at the moment