Fix emojis
In case the comments load before the DOM is ready (which rarely happens, but you never know..) and the extension doesn't notice that the comments have loaded, the extension tries to run the emoji script as soon as it can (and then again, when the comments load).

The first time it ran, it did so without knowing if there actually *was* a comments section, because it was placed outside of the IF statement responsible for verifying this.

This no longer is the case, and so the emoji script should never throw an error, and shouldn't run unless it's on a page with comments.