Fix bugs, move styles to CSS file.
Instead of having the CSS in the JavaScript file (using the `.style` property of an element in the DOM), it is now in inject-style.css. It functions identically.

Additionally, when you collapse an FAQ question, it no longer leaves behind a thin line where the answer was -- the element is fully removed.

The padding on the answer is increased too.

When you click on the answer, the question isn't collapsed, to be consistent with the cursor displayed (the normal one, not "pointer", meaning the area isn't clickable).